Make SafetyRegulatoryInfo target the current API level.
A hardcoded <uses-sdk> element was unnecessarily set in the manifest.
The minSdkVersion was redundant and is overwritten by the build system.
The targetSdkVersion needed to be manually changed each year; by
removing it the app will instead always target 'current' (eg. Tiramisu
at present).

Also fixes an incorrectly placed android:exported attribute which should
have been on the <activity> element, not the <application>.

Bug: 227293013
Test: Manual
Change-Id: Iffe4bdd99916728a0e93e2c0737effd2184421e8